requirements.txt
## ============================================================
## requirements.txt -- Industrial Anomaly Detection (IAD)
## Python 3.10 | CUDA 12.1 | GTX 1660 Ti
## ============================================================

## Core: PyTorch with CUDA 12.1 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121

## Computer Vision & ML
scikit-learn>=1.3.0
matplotlib>=3.7.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

## Data & Config
pyyaml>=6.0
pandas>=2.0.0

## Web App
flask>=3.0.0
python-dotenv>=1.0.0

## XAI Chatbot (Gemini)
google-generativeai>=0.5.0

## Object Detection
ultralytics>=8.0.0

## Model Persistence
joblib>=1.3.0

## Testing
pytest>=7.0.0

## Optional: ONNX Export (not required for web app)
## Uncomment to enable ONNX/TensorRT deployment pipeline:
## onnx>=1.15.0
## onnxruntime-gpu>=1.17.0
## onnxscript>=0.1.0
